requirements.txt
streamlit 
requests 
matplotlib 
fpdf 
crewai 
python-dotenv
mistralai
pandas